requirements.txt
discord.py==2.0.1
python-dotenv==1.0.0
requests==2.31.0
logging==0.5.1.2